Include files and a farewell notice.
Thanks to the costs of maintaining this repository, the future of phishing_pot in 2025 is uncertain. There is a strong possibility that the project will become private, ceasing all updates to this public repository. Or, it's possible that the project will be paused indefinitely. If you use phishing_pot in any way in your company, please consider making a donation to the address below:

bc1q7e3mf5nwmjk9sw8thy35s9r7rq27ta7lzj0d7l

Any help is welcome, thank you.